Evaluation context provider from Component [ANT-3750] (#3043)
Added a new class "EvaluationContextProvider" that allows to easily
generate an EvaluationContext for a component.
ScenarioGroupRepository and LinearProblemData should be given to this
class' ctor ; thus they are considered constant during the simulation.
WARNING : this is a change from before, since LinearProblemData has been
an argument of LinearProblemFiller::addVariables / Constraints /
Objective.
However, given our usage of LinearProblemData, I think we can consider
it constant from now on (hence the TODO in the code)